requirements.txt
numpy
pytorch
tensorflow
pandas